fix(api): move test logic to test files
Ref. metriport/metriport-internal#187
