+++ packages/api/src/command/medical/__tests__/delete-facility.ts
