+++ packages/api/src/command/medical/organization/get-organization.ts
@@ -15,15 +15,3 @@ export const getOrganizationOrFail = async (filter: Filter): Promise<Organizatio
   if (!org) throw new NotFoundError(`Could not find organization`);
   return org;
 };
-
-/**
- * For E2E testing locally and staging.
- * Need to get specific org to increment org number.
- */
-export const getOrganizationById = async (id: string): Promise<OrganizationModel> => {
-  const org = await OrganizationModel.findOne({
-    where: { id },
-  });
-  if (!org) throw new NotFoundError(`Could not find organization`);
-  return org;
-};
