+++ packages/api/src/external/commonwell/__tests__/organization.ts
@@ -1,0 +1,33 @@
+import { Organization as CWOrganization } from "@metriport/commonwell-sdk";
+import { Config } from "../../../shared/config";
+import { capture } from "../../../shared/notifications";
+import { OID_PREFIX } from "../../../shared/oid";
+import { Util } from "../../../shared/util";
+import { makeCommonWellAPI, metriportQueryMeta } from "../api";
+
+const commonWell = makeCommonWellAPI(Config.getMetriportOrgName(), Config.getMemberManagementOID());
+
+/**
+ * For E2E testing locally and staging.
+ */
+export const getOne = async (orgOid: string): Promise<CWOrganization | undefined> => {
+  const { log, debug } = Util.out(`CW get - id ${orgOid}`);
+  const cwId = OID_PREFIX.concat(orgOid);
+  try {
+    const resp = await commonWell.getOneOrg(metriportQueryMeta, cwId);
+    debug(`resp: ${JSON.stringify(resp, null, 2)}`);
+    return resp;
+  } catch (error) {
+    const msg = `[E2E]: Failure getting Org @ CW`;
+    log(msg, error);
+    capture.message(msg, {
+      extra: {
+        orgOid,
+        cwId,
+        cwReference: commonWell.lastReferenceHeader,
+        context: `cw.org.get`,
+      },
+    });
+    throw error;
+  }
+};
