+++ packages/api/src/routes/medical/__tests__/mapi/e2e/mapi.dev.test.e2e.ts
@@ -22,6 +22,8 @@ import { createConsolidated } from "./consolidated";
 import { setupE2ETest, retryFunction, cleanUpE2ETest } from "./shared";
 import { Util } from "../../../../../shared/util";
 import { Config } from "../../../../../shared/config";
+import { getOne as cwOrgGetOne } from "../../../../../external/commonwell/__tests__/organization";
+import { getOne as cwPatientGetOne } from "../../../../../external/commonwell/__tests__/patient";
 
 // Cant delete cw org so we need a test org to increment the id to avoid conflicts
 const INCREMENT_ORG_ID = process.env.TEST_ORG_ID || "";
@@ -63,10 +65,7 @@ if (Config.isStaging() || !Config.isCloudEnv()) {
 
       const fhirOrg = await fhirApi.readResource(ResourceType.Organization, org.id);
 
-      const cwOrg = await retryFunction(
-        async () => await cwCommands.organization.getOne(org.oid),
-        maxRetries
-      );
+      const cwOrg = await retryFunction(async () => await cwOrgGetOne(org.oid), maxRetries);
 
       validateLocalOrg(org, createOrg);
       validateFhirOrg(fhirOrg, createOrg);
@@ -93,7 +92,7 @@ if (Config.isStaging() || !Config.isCloudEnv()) {
       const fhirOrg: FhirOrg = await fhirApi.readResource(ResourceType.Organization, updatedOrg.id);
 
       const cwOrg: CWOrganization | undefined = await retryFunction<CWOrganization | undefined>(
-        async () => await cwCommands.organization.getOne(updatedOrg.oid),
+        async () => await cwOrgGetOne(updatedOrg.oid),
         maxRetries,
         result => result?.name === newName
       );
@@ -158,7 +157,7 @@ if (Config.isStaging() || !Config.isCloudEnv()) {
       const cwPatientId = localPatientResp.data.data.externalData["COMMONWELL"].patientId;
 
       const cwPatient = await retryFunction(
-        async () => await cwCommands.patient.getOne(org, facility, cwPatientId),
+        async () => await cwPatientGetOne(org, facility, cwPatientId),
         maxRetries
       );
 
@@ -200,7 +199,7 @@ if (Config.isStaging() || !Config.isCloudEnv()) {
       const cwPatientId = localPatientResp.data.data.externalData["COMMONWELL"].patientId;
 
       const cwPatient = await retryFunction(
-        async () => await cwCommands.patient.getOne(org, facility, cwPatientId),
+        async () => await cwPatientGetOne(org, facility, cwPatientId),
         maxRetries
       );
