+++ packages/api/src/routes/medical/__tests__/mapi/e2e/mapi.prod.test.e2e.ts
@@ -4,10 +4,10 @@ dotenv.config();
 import { MetriportMedicalApi } from "@metriport/api-sdk";
 import { FhirClient } from "../../../../../external/fhir/api/api";
 import { ResourceType } from "../../../../../external/fhir/shared";
-import cwCommands from "../../../../../external/commonwell";
 import { validateCWOrg, validateFhirOrg, validateLocalOrg } from "./organization";
 import { setupE2ETest } from "./shared";
 import { Config } from "../../../../../shared/config";
+import { getOne as cwOrgGetOne } from "../../../../../external/commonwell/__tests__/organization";
 
 jest.setTimeout(30000);
 
@@ -28,7 +28,7 @@ if (Config.isProdEnv()) {
 
       if (org) {
         const fhirOrg = await fhirApi.readResource(ResourceType.Organization, org.id);
-        const cwOrg = await cwCommands.organization.getOne(org.oid);
+        const cwOrg = await cwOrgGetOne(org.oid);
 
         validateLocalOrg(org);
         validateCWOrg(cwOrg);
